requirements.txt
aiogram
aiohttp
pytesseract
Pillow
